+++ src/commands/serve.js
@@ -1,7 +1,8 @@
 const { Command, flags } = require('@oclif/command')
 const { exec } = require('child_process')
+const Build = require('./build')
 const Cache = require('./cache')
-const nodemon = require('nodemon');
+const chokidar = require('chokidar')
 
 class ServeCommand extends Command {
   async run() {
@@ -14,36 +15,29 @@ class ServeCommand extends Command {
     process.env.FORMIDABLE_PORT = flags.port;
 
     if (flags.dev) {
-      const minify     = flags.minify ? '--minify' : '';
-      const _minify    = flags['no-minify'] ? '--no-minify' : '';
-      const _sourceMap = flags['no-sourcemap'] ? '-S' : '';
-      const _hashing   = flags['no-hashing'] ? '-H' : '';
-      const silent     = flags.debug ? '' : '--silent';
+      await Build.run(['--env', flags.env, '--continue']);
 
-      const server = nodemon({
-        ext: 'imba',
-        ignore: ['dist', '.formidable'],
-        exec: `node node_modules/.bin/craftsman build ${silent} ${_minify} ${minify} ${_sourceMap} ${_hashing} && node node_modules/.bin/imba server.imba`
+      const watcher = chokidar.watch('./**/*.imba', {
+        ignored: ['.formidable', 'node_modules', 'public', 'dist', 'test', 'tests'],
+        persistent: true,
+        ignoreInitial: true
       });
 
-      server.on('start', () => {
-        console.log("\x1b[32mStarting Formidable development server\x1b[0m");
-      }).on('crash', () => {
-        console.log('Application crashed');
-      }).on('restart', () => {
-        console.log("\x1b[32mRestarting Formidable development server\x1b[0m");
-      });
+      watcher.on('all', async () => await Build.run(['--env', flags.env, '--continue']));
 
-      process.on('SIGINT', function() {
+      process.on('SIGINT', () => {
+        watcher.close();
         process.exit();
       });
-
-      return;
     }
 
-    await Cache.run(['--env', flags.env, '--continue']);
+    if (!flags.dev) {
+      await Cache.run(['--env', flags.env, '--continue']);
+    }
 
-    const command = `node node_modules/.bin/imba server.imba`;
+    const watch = flags.dev ? '-w' : '';
+
+    const command = `node node_modules/.bin/imba ${watch} server.imba`;
 
     const server = exec(command);
 
@@ -63,7 +57,7 @@ ServeCommand.flags = {
   'no-hashing': flags.boolean({char: 'H', description: 'Disable hashing' }),
   'no-minify': flags.boolean({char: 'M', description: 'Disable minifying', default: true }),
   'no-sourcemap': flags.boolean({char: 'S', description: 'Disable sourcemaps', default: true }),
-  dev : flags.boolean({char: 'd', description: 'Serve in dev mode'}),
+  dev : flags.boolean({char: 'd', description: 'Serve in dev mode (build, serve and watch)'}),
   env: flags.option({ char: 'e', description: 'The environment to build for', default: 'local', options: ['local', 'testing', 'development', 'staging', 'production'] }),
   exit: flags.boolean({char: 'e', description: 'Exit after cache', default: false }),
   minify: flags.boolean({char: 'm', description: 'Minify generated files'}),